Refactor flv-test to expose reporting
