+++ tests/runner/src/utils/test_meta/environment.rs
@@ -1,5 +1,5 @@
 use crate::setup::environment::{EnvironmentType};
-use crate::test_runner::FluvioTestMeta;
+use crate::test_runner::fluvio_test_meta::FluvioTestMeta;
 use structopt::StructOpt;
 use std::fmt::Debug;
 use std::num::ParseIntError;
