+++ tests/runner/src/utils/test_runner/mod.rs
@@ -1,6 +1,8 @@
+pub mod fluvio_test_meta;
+
 #[allow(unused_imports)]
 use fluvio_command::CommandExt;
-use crate::test_meta::{TestCase, TestOption, TestResult};
+use crate::test_meta::{TestCase, test_result::TestResult};
 use crate::test_meta::environment::{EnvDetail, EnvironmentSetup};
 use crate::test_meta::derive_attr::TestRequirements;
 use fluvio::{Fluvio, FluvioError};
@@ -9,7 +11,6 @@ use fluvio::metadata::topic::TopicSpec;
 use hdrhistogram::Histogram;
 use fluvio::{TopicProducer, RecordKey, PartitionConsumer};
 use std::time::Duration;
-use async_lock::RwLock;
 use futures_lite::stream::StreamExt;
 use tracing::debug;
 use fluvio::Offset;
@@ -30,8 +31,32 @@ pub struct FluvioTestDriver {
 }
 
 impl FluvioTestDriver {
+    pub fn new(client: Arc<Fluvio>) -> Self {
+        Self {
+            client,
+            num_topics: 0,
+            num_producers: 0,
+            num_consumers: 0,
+            bytes_produced: 0,
+            bytes_consumed: 0,
+            produce_latency: Histogram::<u64>::new_with_bounds(1, u64::MAX, 2).unwrap(),
+            consume_latency: Histogram::<u64>::new_with_bounds(1, u64::MAX, 2).unwrap(),
+            topic_create_latency: Histogram::<u64>::new_with_bounds(1, u64::MAX, 2).unwrap(),
+        }
+    }
+
     pub fn get_results(&self) -> TestResult {
-        TestResult::default()
+        TestResult {
+            num_topics: self.num_topics as u64,
+            num_producers: self.num_producers as u64,
+            num_consumers: self.num_consumers as u64,
+            bytes_produced: self.bytes_produced as u64,
+            bytes_consumed: self.bytes_consumed as u64,
+            topic_create_latency: self.topic_create_latency.value_at_percentile(99.0),
+            produce_latency: self.produce_latency.value_at_percentile(99.0),
+            consume_latency: self.consume_latency.value_at_percentile(99.0),
+            ..Default::default()
+        }
     }
 
     // Wrapper to getting a producer. We keep track of the number of producers we create
@@ -189,48 +214,5 @@ impl FluvioTestDriver {
         }
 
         true
-    }
-}
-
-#[derive(Debug)]
-pub struct FluvioTestMeta {
-    pub name: String,
-    pub test_fn: fn(Arc<RwLock<FluvioTestDriver>>, TestCase) -> Result<TestResult, TestResult>,
-    pub validate_fn: fn(Vec<String>) -> Box<dyn TestOption>,
-    pub requirements: fn() -> TestRequirements,
-}
-
-inventory::collect!(FluvioTestMeta);
-
-impl FluvioTestMeta {
-    pub fn all_test_names() -> Vec<&'static str> {
-        inventory::iter::<Self>
-            .into_iter()
-            .map(|x| x.name.as_str())
-            .collect::<Vec<&str>>()
-    }
-
-    pub fn from_name<S: AsRef<str>>(test_name: S) -> Option<&'static Self> {
-        inventory::iter::<Self>
-            .into_iter()
-            .find(|t| t.name == test_name.as_ref())
-    }
-
-    pub fn set_topic(test_reqs: &TestRequirements, test_case: &mut TestCase) {
-        if let Some(topic) = &test_reqs.topic {
-            test_case.environment.set_topic_name(topic.to_string());
-        }
-    }
-
-    pub fn set_timeout(test_reqs: &TestRequirements, test_case: &mut TestCase) {
-        // Set timer
-        if let Some(timeout) = test_reqs.timeout {
-            test_case.environment.set_timeout(timeout)
-        }
-    }
-
-    pub fn customize_test(test_reqs: &TestRequirements, test_case: &mut TestCase) {
-        Self::set_topic(test_reqs, test_case);
-        Self::set_timeout(test_reqs, test_case);
     }
 }
